+++ src/aac_metrics/classes/base.py
@@ -12,13 +12,16 @@ try:
 except ModuleNotFoundError:
 
     class Metric(nn.Module):
+        """Base Metric module used when torchmetrics is not installed."""
+
         full_state_update: Optional[bool] = False
         higher_is_better: Optional[bool] = None
         is_differentiable: Optional[bool] = False
 
+        # The minimal value of the main global score of the metric.
         min_value: Optional[float] = None
+        # The maximal value of the main global score of the metric.
         max_value: Optional[float] = None
-        is_linear: Optional[bool] = None
 
         def compute(self) -> Any:
             return None