+++ src/aac_metrics/classes/evaluate.py
@@ -15,6 +15,11 @@ logger = logging.getLogger(__name__)
 
 
 class CustomEvaluate(Metric, list[Metric]):
+    """Evaluate candidates with multiple references with custom metrics.
+
+    For more information, see :func:`~aac_metrics.functional.evaluate.custom_evaluate`.
+    """
+
     full_state_update = False
     higher_is_better = None
     is_differentiable = False
@@ -22,27 +27,27 @@ class CustomEvaluate(Metric, list[Metric]):
     def __init__(
         self,
         use_ptb_tokenizer: bool = True,
+        cache_path: str = "$HOME/aac-metrics-cache",
         java_path: str = "java",
         tmp_path: str = "/tmp",
-        cache_path: str = "$HOME/aac-metrics-cache",
         verbose: int = 0,
         metrics: Union[str, Iterable[Metric]] = "all",
     ) -> None:
         if isinstance(metrics, str):
             metrics = _get_metrics_list(
                 metrics,
+                cache_path=cache_path,
                 java_path=java_path,
                 tmp_path=tmp_path,
-                cache_path=cache_path,
                 verbose=verbose,
             )
 
         Metric.__init__(self)
         list.__init__(self, metrics)
         self._use_ptb_tokenizer = use_ptb_tokenizer
+        self._cache_path = cache_path
         self._java_path = java_path
         self._tmp_path = tmp_path
-        self._cache_path = cache_path
         self._verbose = verbose
 
         self._candidates = []
@@ -54,9 +59,9 @@ class CustomEvaluate(Metric, list[Metric]):
             self._mult_references,
             self._use_ptb_tokenizer,
             self,
+            cache_path=self._cache_path,
             java_path=self._java_path,
             tmp_path=self._tmp_path,
-            cache_path=self._cache_path,
             verbose=self._verbose,
         )
 
@@ -75,19 +80,24 @@ class CustomEvaluate(Metric, list[Metric]):
 
 
 class AACEvaluate(CustomEvaluate):
+    """Evaluate candidates with multiple references with all Audio Captioning metrics.
+
+    For more information, see :func:`~aac_metrics.functional.evaluate.aac_evaluate`.
+    """
+
     def __init__(
         self,
         use_ptb_tokenizer: bool = True,
+        cache_path: str = "$HOME/aac-metrics-cache",
         java_path: str = "java",
         tmp_path: str = "/tmp",
-        cache_path: str = "$HOME/aac-metrics-cache",
         verbose: int = 0,
     ) -> None:
         super().__init__(
             use_ptb_tokenizer,
+            cache_path,
             java_path,
             tmp_path,
-            cache_path,
             verbose,
             "aac",
         )