+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] 2022-09-30
+### Added
+- Documentation for metric functions and classes.
+- A second larger example for unit testing.
+
+### Changed
+- Update README information, references and description.
+
+### Fixed
+- SPIDEr-max computation with correct global and local outputs.
+- Unit testing for computing SPICE metric from caption-evaluation-tools.
+
 ## [0.1.0] 2022-09-28
 ### Added
 - BLEU, METEOR, ROUGE-l, SPICE, CIDEr and SPIDEr metrics functions and modules.